move scientific notation rounding to rounding.py
