+++ safe/utilities/rounding.py
@@ -1,7 +1,6 @@
 # coding=utf-8
 """Rounding and number formatting."""
-
-
+from decimal import Decimal
 from math import ceil
 
 from safe.definitions.units import unit_mapping
@@ -253,3 +252,24 @@ def fatalities_range(number):
             return range_format.format(
                 min_range=add_separators(min_range),
                 max_range=add_separators(max_range))
+
+
+def html_scientific_notation_rate(rate):
+    """Helper for convert decimal rate using scientific notation.
+
+    For example we want to show the very detail value of fatality rate
+    because it might be a very small number.
+
+    :param rate: Rate value
+    :type rate: float
+
+    :return: Rate value with html tag to show the exponent
+    :rtype: str
+    """
+    rate = '%.2E' % Decimal(rate)
+    html_rate = rate.split('E')
+    # we use html tag to show exponent
+    html_rate[1] = '10<sup>{exponent}</sup>'.format(
+        exponent=html_rate[1])
+    html_rate.insert(1, 'x')
+    return ''.join(html_rate)
